Add quantr errors to Circuit::add_gates
Adds extra checking to the user input of Circuit::add_gates,
specifically for gates with a/multiple control node/nodes.

The method already verifies if the given vector is of a correct size.

Now, this will push multi-controlled gates to another column, so that
they are isolated. By doing so, this ensures that no other
nodes/positions of gates overlap with other gates' nodes-  reducing
processing time as there are less checks - and it makes for easier
printing for the Printer module, that is the assumption of isolated
multi-controlled gates drastically simplifies printing the diagram as
there are fewer checks to make.

For every multi-controlled gate, it will check if the nodes are
overlapping with themselves, or with the gate head itself (the target).

Combining the two methods above, that is isolating these gates and then
checking each one for overlapping positions, ensures that the entire
circuit will have no gate overlapping features.

The single gates cannot overlap due to the intrinsic nature of requiring
a vector of gates as input.

Circuit::add_repeating_gate uses Circuit::add_gates, so error
implementation will also be included in for this method.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>