Add maximum circuit size and remove function
Implemented a maximum circuit size (for now declared as a usize constant
of 50 in circuit.rs), that is guaranteed to be adhered to upon
initlialisation of the circuit. This has resulted in a Result being
returned upon initialisation, ensuring that the declared circuit is
valid. ** This induces a breaking change. ** The tests have been updated
to unwrap the declared circuit.

The method Circuit::add_repeating_gate used a duplicate function in
finding unique values. So, removed this unnecersary function, and reused
the function that is used for Circuit::add_gates.

The algorithm to find unique values is O(n), however requires extra
memory as it initialises a boolean list with the same size of the number
of wires (the new usize constant). This list will be removed from the
stack at the end of the algorithm, so memory is still minimised.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>