Convert public to private methods and add const
Added a ZERO_MARGIN const to circuit, that is used to declare when a
state is considered to have zero amplitude. This is useful for
optimising the simulator, as near zero states can be disregarded.

Some public methods in states.rs have been converted to private. There
is a fine balance between allowing the user to access methods that
should include proepr error handling, and only allowing quantr to access
these methods and in doing so, removing error checks to increase speed.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>