Make public methods private
An overhaul of making methods available to the user and changing the
visibility of error::QuantrError.

First, error.rs has been moved inline with circuit.rs, and QuantrError
has been made unavailable to the user. Changes to printer.rs reflect
this change.

Secondly, a collection of public methods have been converted to
pub(super) so that the Circuit struct can still use them, but cannot be
used by the user. In doing so, input validation does not need to be
included, as Circuit can guarantee valid input, which results in faster
simualtion as no unnecersary input validation has to be made. If these
methods were exposed to the user, input validation should be
implemented.

The methods in SuperPosition that remain visible for the user are mainly
for setting superpositions and getting amplitudes based on list
positions and product states, that all have input validation.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>