Refractor in Circuit.rs and edit docs
A small re-shuffling of methods in Circuit.rs. This is mainly for
presenting the methods in a logical fashion in the documentation. That
is, the methods are presented in increasing order of complexity, and to
which they might be used in an algorithm.

Some documentation was edited, removing the reference to the struct
QuantrError.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>