Add error checking to methods
From respect to the pull request, this has added complete error checking
to the method that creates a superposition from a HashMap of
ProductStates. There remains another version that is used in the main
algorithm, which does not implement error checking. It has `_unchecked`
at the end, to denote this. This can be done safely as the main algorthm
can guarantee valid input.

The quick start guide has also been edited as breaking changes were made
to the Circuit, where it has to be unwrapped. In this wrap, it will also
check that there is at least one wire.

Again, this implements breaking changes.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>