Correct SuperPosition::as_hash_map
The bin count of repeated measurements was only returning states that
were vanishing... and those states are not included in a hash map. So,
the hash map returned was empty. This has now been corrected, and now a
test has been included in grovers to ensure that a sensible bin count is
returned.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>