Fix printer and pushing multi gates
The printer will now adjust for variable name length of gates. Moreover,
a bug in the pushing of multi-controlled custom gates is fixed; that is
they now move to a new column.

The printer before adjusted for single gates of gate names longer than
one length. However, the printer will now do this for double, triple and
custom gates with multiple nodes. Custom gates will always be treated as
multi gates, even if they have one node.

The functions in the printer module and it's printer struct are now
cleaner and better 'programmed' than before, making more use of structs
and Options. The module should be more maintainable and easier to
understand.

Lastly, there was a bug fixed that didn't push multi-control gates to
isolated columns; it does now.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>