Add ability to convert product state to superpos
A product state struct can now be converted to a superposition struct
simply by calling it's method `to_super_position()`.

This allows for easier manipulation in constructing custom functions.
For example, if a product state is altered by flipping a binary digit,
then the custom function needs to return a superposition. This now makes
it far easier to do.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>