Add warning to non-ASCII custom name
Non-ascii characters cannot guarantee the length and display length to
be equal. Thus, this may cause issue for the printer when displaying the
circuit diagram, that takes this into account.

A warning has been added pointing out this issue, in addition to it
being promoted to a quantr error in the next breaking update.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>