Re-arrange access points to structs and modules
The structs can now be accessed directly through `quantr::{StandardGate,
Circuit,..., states::{ProductState}}`. The only module which is
accessible is `states`.

There has been no re-arranging of the project strucutre itself, and all
tests, and unit tests, have been updated to adhere to the new accessing
of structs and module.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>