Add chains of gate additions and changed inputs
The user can now chain methods that add gates to the circuit. Moreover,
all `Vec` inputs of these methods have been changed to slices.

All tests have been updated, otherwise they break. The tests and
examples have been updated to include the chaining method calls. That
is, a mutable reference to the circuit is passed down a chain of
methods.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>