Make cargo fmt skip tests and examples
Made sure that cargo fmt doesn't affect the tests, unit tests and
examples as it makes all the .unwrap() take a new line. I prefer the
syntac of the unwrap() not taking a new line.

This may be reviewed later, but it's benifical to have unwrap() as it
cleary highlights the line to which there was a mistake.

Ran cargo fmt.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>