Make QuantrError public and tidy examples
Made QuantrError public so it can make coding in main functions neater
and more readable when the return signature is Result<(), QuantrError>.

The examples showcase this idea, however the tests and code snippet
examples within the source code itself, do not use ? instead of
unwrap(). This is to show explicitly how these methods return Results.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>