Refractor and add control node check
The enum `StandardGate` was refractored to `Gate`. Makes it more
manegavle to write out; there was no need for the Standard prefix.

The check for the control nodes on the gates were missing for the newly
implemented gates. This has now been corrected, with an additional test
for making sure control nodes are not greater than the number of qubits
in the circuit.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>