Fix qft example and ASCII error
The `qft.rs` in the examples folder has been fixed and is now correctly
working. This is a good demonstration how mini circuits can be
constructed to define custom gates. A warning has also been added to the
custom function doc that the user themselves have to ensure the
unitarity of the function.

The ASCII warning of using non-ASCII strings to name custom gates has
been bumped up to a warning. If the user uses utf-8, this most likely
results in the diagram being completely difigured and unrecoverable.
This is ude to the visual width of Utf-8 characters not being their
'actual width', which messes up the printer.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>