Renovate register changing
The method that was primiarly used if there was a need for change of
register was one that basically copied the simualte function. Now, the
circuit struct has been changed so it has the field
`Option<SuperPosition>`, which is edited with
`Circuit::change_register`. This has cleared up a lot of the code.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>