Review documentation
Mostly grammatical changes in code documentation, such as grammar and
deleting sentences that reference now delted functions for version
0.3.0.

Added a qft test in the "tests" folder, as the output register is quite
unique. Have change one function, the ProductState::new to now return a
result as an error could be passing it an empty qubit slice.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>