Restructure code in circuit and states
The state structs (qubit, superposition and product state), have been
organised into separate files. The tests have also moved into filed
which test the struct. This will hopefully compartmentalise the code and
will make it easier to manage.

The circuit.rs file has been broken down into itself and a
circuit/simulation.rs file, where the later contains most of the
functions that are used to siulate the circuit. The former file has all
the public functions, in addition to the private ones that checks the
user input.

Finally, the Quick Start Guide has been fixed, in that the vec! macro
has been removed as inputs to the add gate functions.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>