Add Option return to custom func variant
The idea is that when a custom function returns a Option<SuperPosition>,
if it's some then it means that the state that has been inserted has
changed. If None, then the inputed state has not been changed and thus
bulk of the operation can be skipped. This is particulary useful for
large CNOT operaters. This will hopefully increase the speed of applying
large custom gates.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>