Implement From traits for states and super pos
This replaces the `into` methods that were generated for Qubit and
ProductState structs with the `From` trait implementations. This is to
be more inline with the Rust library.

Moreover, a new file was generated to hold the SuperPositionIterator.
This tidies up the source code.

The ProductState::invert_digit now returns a mutable reference of the
product state that it altered.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>