Add documentation to states
Lots of documentation added to super position and prouct states, all
inclduing examples. This drive should now make quantr more accesible.

Moreover, the fields of the Product and superpositions have been made
private, so the user has to directly use the methods (which contains
user validation) to initialise any of these objects. This now further
guarantees that all states that the user initialises will be able to be
processed without error.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>