Restructre code for better maintainability
The `GateSize` enum has been replaced with `GateCategories` that makes
better use of match statements, by allowing a completely exhaustive
pattern. This has reduced code massively, in addition to reducing the
number of panic statements and unnecessary branches in the match
statements.

The GateSize was used in the printer struct, however that's been removed
too. More code cleaning needs to be done for this file, for
maintainability reasons, but this will be left for a future update.

Lastly, most functions that linked the gates to functions, names or
returning control nodes have been moved to `gates.rs`.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>