Implement SimulatedCircuit and update unit tests
Added SimulatedCircuit struct, which is produced when the Circuit struct
is simulated. More functionality needs to be added to SimulatedCircuit,
for instance a partial simulation method for mid-circuit measurements.

The unit tests and examples have been updated to use the
SimulatedCircuit struct, so now all tests and examples pass.

The purpose for adding this extra struct is to increase the saftey of
interfacing with quantr; with SimulatedCircuit struct you are guaranteed
to have a circuit with a superposition that can be measured, rather than
a broken circuit.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>