Update handling of circuit register
Previously, there was cloning of the superposition if the register was
changed, leading to unnecersary memory consumption. This has no been
changed, where only one superposition is kept within the struct, which
is used fir changing the register, and simulation.

The internal simulation methods were updated, to better handle the
Option of Superposition too (no .unwraps() were necersary!).

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>