Update CHANGELOG and better observed enums
Updated the CHANGELOG.md with the recent changes made for 0.6.0 so far.

Added some methods to the Measurement enum so that the value it wraps
can be moved outside. Thus, the user can avoid the explicit handling of
unwrapping the value through `let` statements if they want to. A similar
take method was added to the `SimulatedCircuit` so that the
superpositiion could be moved out. This resulted in the simplification
of many custom gates in the examples where cloning was being used.
Clones have no been removed.

Tests and examples have been updated accordingly.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>