Add measurement ability to SuperPosition
A method was added that acts as a measurement of a Superposition, and
returns the state that the Superposition collapsed to.

In turn, this has internally improved the readability of the
repeat_measurement code in Simulated Circuit. However, as Custom gates
do not have to be Unitary, this has left the possibility that registers
may not conserve probability. Therefore, the repeat simulation methods
will no send a warning to stderr, alerting if a state did not collape to
anything.

All Quantr Warnings now print to stderr, instead of stdout.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>