Add disable warning feature and refactor methods
Added the ability to turn off warnings for SimulatedCircuit and Printer.

The measuring methods, get_superposition and repeat_measurement, have
been renamed to get_state and measure_all respectively. This are more
succicent. The docs have been updated accordingly, in addition to the
location of methods reshuffled so the important ones appear first for
the user.

A warning will now appear is the user is measuring states repeadetly
using SimulatedCircuit::measure_all with Custom Gates present; as this
may lead to incorrect results. The warning suggests to toggle the
cache ability off. This will simulate properly (for instance if a Custom
gate implements a mid-circuit measurement).

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>