Add register cache warnings for measuring
The measure_all function, when there are custom gates detected in the
circuit (as all other gates lead to unitary mappings and so will be safe
to produce measurements), it will warn the user that register is cached
and may lead to run results, as mixed states can be introduced. An
example will be given in a future commit; where one has to use the
measure_all_without_cache method.

The From<Qubit> trait for Superposition was implemented.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>