Implement post selection example
The main algorithm was overhauled to allow some quantum channels to be
implemented. THe main algorithm, that is calculating how gates operate
on the state vectors have been changed so that custom gates not
conserving probability can work. Before, the algorithm was assuming that
all probabilities were conserved, and hence could use a shortcut. Now,
the algorithm will most likely take longer to execute.

Some methods, that allow the creation of non-probability conserving
statevectors, have been made public to the user. This is required to
implement arbitary quantum channels, for example pos selection (of
which) there is an example in the `examples` folder.

Signed-off-by: Andrew Barlow <a.barlow.dev@gmail.com>